Moved mathematical functions from inversepowerlaw.py to a separate math.py script
